Uploaded Initial Rights & Responsibilities Doc
--- This is version 1 and we understand that this needs more work, better definition and to grow with our community. ---

Accessibility means your website can reach more people, be compliant with laws, and overall meet the new industry standards, but for some teams that can seem overwhelming. It does not have to be. Your groups likely already have the resources you need, it’s a matter of organization and understanding who is responsible for which parts of the process. In this guide we will breakdown the roles and responsibilities of both the design/development team and that of the content authors/managers. By outlining how each sub-group can make a positive impact in accessibility, and listing out some of the tools to aid them on their way, this once overwhelming initiative can become your standard operating procedure.  
 
It’s about simplifying, understanding and practice.